students/IS22/Bithor Leonid/lab2/code/requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cheroot==11.1.2
cotyledon==2.1.0
croniter==6.0.0
cryptography==46.0.3
debtcollector==3.0.0
decorator==5.2.1
dnspython==2.8.0
dogpile.cache==1.5.0
eval_type_backport==0.3.1
eventlet==0.40.4
exceptiongroup==1.3.1
fasteners==0.20
frozenlist==1.8.0
futurist==3.2.1
googleapis-common-protos==1.72.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
invoke==2.2.1
iso8601==2.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keystoneauth1==5.12.0
keystonemiddleware==10.12.0
kombu==5.6.1
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
mistral==21.0.0
mistral-lib==3.5.0
mistralai==1.10.0
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
netaddr==1.3.0
networkx==3.4.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
os-service-types==1.8.2
oslo.cache==4.0.0
oslo.concurrency==7.2.0
oslo.config==10.1.0
oslo.context==6.2.0
oslo.db==17.4.0
oslo.i18n==6.7.1
oslo.log==8.0.0
oslo.messaging==17.1.0
oslo.metrics==0.14.0
oslo.middleware==7.0.0
oslo.policy==4.7.0
oslo.serialization==5.8.0
oslo.service==4.4.1
oslo.utils==9.2.0
osprofiler==4.3.0
packaging==25.0
paramiko==4.0.0
Paste==3.10.1
PasteDeploy==3.1.0
pbr==7.0.3
pecan==1.7.0
prettytable==3.17.0
prometheus_client==0.23.1
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pycadf==4.0.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
PyNaCl==1.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-keystoneclient==5.7.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
repoze.lru==0.7
requests==2.32.5
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.30.0
simplegeneric==0.8.1
six==1.17.0
SQLAlchemy==2.0.45
statsd==4.0.1
stevedore==5.6.0
tenacity==9.1.2
testresources==2.0.2
testscenarios==0.5.0
testtools==2.8.2
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
WebOb==1.8.9
wrapt==2.0.1
WSME==0.12.1
yappi==1.7.3
yaql==3.2.0
yarl==1.22.0
zipp==3.23.0
